fix long totalAmount with float

+++ app/src/main/java/tr/edu/iyte/installmentor/database/entities/Product.java
@@ -9,11 +9,11 @@ public class Product extends Entity {
     private float totalAmount;
     private float remainingAmount;
 
-    public Product(long cardId, long totalAmount, String description, Date buyDate) {
+    public Product(long cardId, float totalAmount, String description, Date buyDate) {
         this(0, cardId, totalAmount, description, buyDate);
     }
 
-    public Product(long id, long cardId, long totalAmount, String description, Date buyDate) {
+    public Product(long id, long cardId, float totalAmount, String description, Date buyDate) {
         this.id = id;
         this.cardId = cardId;
         setTotalAmount(totalAmount);
